Cleanup: various non functional changes
- Use const arguments.
- Quiet unused assignment warnings.
- Use explicit cast to suppress CPPCHECK's truncLongCastReturn warning.

+++ source/blender/blenlib/BLI_hash.hh
@@ -144,7 +144,8 @@ TRIVIAL_DEFAULT_INT_HASH(uint64_t);
 template<> struct DefaultHash<float> {
   uint64_t operator()(float value) const
   {
-    return *reinterpret_cast<uint32_t *>(&value);
+    /* Explicit `uint64_t` cast to suppress CPPCHECK warning.  */
+    return uint64_t(*reinterpret_cast<uint32_t *>(&value));
   }
 };
